server/pyproject.toml
[tool.poetry]
name = "dates-in-sg"
version = "0.1.0"
description = ""
authors = ["Bryan Ang <bryanang2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
routes = "^2.5.1"
flask = "^3.0.0"
pymongo = "^4.6.1"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
pylint = "^3.0.3"
pytest = "^8.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
